+++ client/CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `alias_id_to_bech32` and `nft_id_to_bech32` moved to `types` crate;
+- `is_address_valid` moved to `types` crate;
 
 ## 2.0.1-rc.7 - 2023-03-09